requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
fuzzywuzzy==0.18.0
itsdangerous==2.2.0
Jinja2==3.1.3
Levenshtein==0.25.1
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pytesseract==0.3.10
python-Levenshtein==0.25.1
rapidfuzz==3.8.1
Werkzeug==3.0.2